# api/openapi.json
{"openapi":"3.1.1","jsonSchemaDialect":"https://json-schema.org/draft/2020-12/schema","info":{"title":"Compute Relay API","version":"0.1.0-alpha.1","description":"Provider-neutral, operator-hosted finite compute job API. This contract is planned and not evidence of a live server or provider capability. Runtime health/info and object handlers are implemented offline as library components; production CLI/SQLite composition and job routes remain planned.","license":{"name":"Apache-2.0"}},"servers":[{"url":"http://127.0.0.1:7331","description":"Default loopback runtime address"}],"security":[{"bearerAuth":[]}],"x-implementation-status":"planned","tags":[{"name":"runtime"},{"name":"jobs"},{"name":"operations"},{"name":"objects"}],"paths":{"/healthz":{"get":{"summary":"Local process liveness","operationId":"getHealth","tags":["runtime"],"security":[],"x-implementation-status":"implemented-offline","responses":{"200":{"description":"Runtime process is alive","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["status"],"properties":{"status":{"const":"ok"}}}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"503":{"$ref":"#/components/responses/Error503WithHeaders"}}}},"/readyz":{"get":{"summary":"Local runtime readiness","operationId":"getReadiness","tags":["runtime"],"x-implementation-status":"implemented-offline","responses":{"200":{"description":"Required local services are ready","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["status"],"properties":{"status":{"const":"ready"}}}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"503":{"$ref":"#/components/responses/Error503WithHeaders"},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"401":{"$ref":"#/components/responses/Error401WithHeaders"}}}},"/v1/info":{"get":{"summary":"Runtime and API version information","operationId":"getRuntimeInfo","tags":["runtime"],"x-implementation-status":"implemented-offline","responses":{"200":{"description":"Public runtime information","content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["api_version","runtime_version","implementation_status","features","job_admission"],"properties":{"api_version":{"const":"compute-connector/v1alpha1"},"runtime_version":{"type":"string","minLength":1},"implementation_status":{"const":"implemented-offline"},"features":{"type":"array","maxItems":3,"uniqueItems":true,"items":{"enum":["workspace_auth","object_upload","object_metadata"]}},"job_admission":{"const":"not_implemented"}}}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"401":{"$ref":"#/components/responses/Error401WithHeaders"},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"503":{"$ref":"#/components/responses/Error503WithHeaders"}}}},"/v1/workspaces/{workspace_id}/jobs/validate":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}}],"post":{"summary":"Validate a job without allocating compute","operationId":"validateJob","x-implementation-status":"planned","responses":{"200":{"description":"Validation result","content":{"application/json":{"schema":{"$ref":"./schemas/job-validation.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"./schemas/job-spec.v1alpha1.schema.json"}}}}}},"/v1/workspaces/{workspace_id}/jobs":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}}],"post":{"summary":"Durably accept an immutable job","operationId":"createJob","x-implementation-status":"planned","responses":{"202":{"description":"Job accepted locally","headers":{"Location":{"description":"Job status URI","schema":{"type":"string","format":"uri-reference"}}},"content":{"application/json":{"schema":{"$ref":"./schemas/job-admission.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}},"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"./schemas/job-spec.v1alpha1.schema.json"}}}},"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":8,"maxLength":256}}]}},"/v1/workspaces/{workspace_id}/jobs/{job_id}":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"job_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/job_id"}}],"get":{"summary":"Read one job and active attempt state","operationId":"getJob","x-implementation-status":"planned","responses":{"200":{"description":"Job status","content":{"application/json":{"schema":{"$ref":"./schemas/job-status.v1alpha1.schema.json"}}}},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"404":{"$ref":"#/components/responses/Error404"}}}},"/v1/workspaces/{workspace_id}/jobs/{job_id}/cancel":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"job_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/job_id"}}],"post":{"summary":"Persist cancellation intent","operationId":"cancelJob","x-implementation-status":"planned","responses":{"202":{"description":"Cancellation operation accepted","content":{"application/json":{"schema":{"$ref":"./schemas/operation.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}}}},"/v1/workspaces/{workspace_id}/jobs/{job_id}/retry":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"job_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/job_id"}}],"post":{"summary":"Create an explicit new compute attempt","operationId":"retryJob","x-implementation-status":"planned","responses":{"202":{"description":"Retry operation accepted","content":{"application/json":{"schema":{"$ref":"./schemas/operation.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}},"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string","minLength":8,"maxLength":256}}]}},"/v1/workspaces/{workspace_id}/jobs/{job_id}/reconcile":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"job_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/job_id"}}],"post":{"summary":"Reconcile existing provider identity without new compute","operationId":"reconcileJob","x-implementation-status":"planned","responses":{"202":{"description":"Reconciliation operation accepted","content":{"application/json":{"schema":{"$ref":"./schemas/operation.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}}}},"/v1/workspaces/{workspace_id}/jobs/{job_id}/collect":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"job_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/job_id"}}],"post":{"summary":"Retry collection without rerunning compute","operationId":"collectJob","x-implementation-status":"planned","responses":{"202":{"description":"Collection operation accepted","content":{"application/json":{"schema":{"$ref":"./schemas/operation.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400"},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"409":{"$ref":"#/components/responses/Error409"},"422":{"$ref":"#/components/responses/Error422"},"429":{"$ref":"#/components/responses/Error429"},"503":{"$ref":"#/components/responses/Error503"}}}},"/v1/workspaces/{workspace_id}/operations/{operation_id}":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"operation_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/operation_id"}}],"get":{"summary":"Read one durable control operation","operationId":"getOperation","x-implementation-status":"planned","responses":{"200":{"description":"Operation status","content":{"application/json":{"schema":{"$ref":"./schemas/operation.v1alpha1.schema.json"}}}},"401":{"$ref":"#/components/responses/Error401"},"403":{"$ref":"#/components/responses/Error403"},"404":{"$ref":"#/components/responses/Error404Operation"}}}},"/v1/workspaces/{workspace_id}/objects":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}}],"post":{"summary":"Stream an immutable object into workspace storage","operationId":"uploadObject","tags":["objects"],"x-implementation-status":"implemented-offline","description":"Requires write scope. Streams raw bytes; no archive extraction or provider staging. Content-Length and X-Content-SHA256 are optional declarations. Unknown-length/chunked and empty objects are accepted within operator limits (2 GiB default). 201 is sent only after verified blob publication and ownership metadata commit; production SQLite composition is deferred. No resumable upload or deduplication promise.","parameters":[{"name":"X-Content-SHA256","in":"header","required":false,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/sha256"}}],"requestBody":{"required":true,"content":{"application/octet-stream":{"schema":{"type":"string","format":"binary"}},"application/gzip":{"schema":{"type":"string","format":"binary"}}}},"responses":{"201":{"description":"Verified bytes and workspace ownership metadata committed","headers":{"Location":{"schema":{"type":"string","format":"uri-reference"},"description":"Workspace-scoped metadata URI"},"X-Request-ID":{"$ref":"#/components/headers/RequestID"}},"content":{"application/json":{"schema":{"$ref":"./schemas/object.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"401":{"$ref":"#/components/responses/Error401WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"408":{"$ref":"#/components/responses/Error408WithHeaders"},"413":{"$ref":"#/components/responses/Error413WithHeaders"},"415":{"$ref":"#/components/responses/Error415WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"503":{"$ref":"#/components/responses/Error503WithHeaders"}}}},"/v1/workspaces/{workspace_id}/objects/{object_id}":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}},{"name":"object_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/object_id"}}],"get":{"summary":"Read committed workspace object metadata","operationId":"getObject","tags":["objects"],"x-implementation-status":"implemented-offline","description":"Requires read scope. Does not expose host paths or provider URLs. An absent or foreign object has the same 404 response. Uncommitted filesystem bytes are not visible here.","responses":{"200":{"description":"Immutable object metadata","content":{"application/json":{"schema":{"$ref":"./schemas/object.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"401":{"$ref":"#/components/responses/Error401WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"404":{"$ref":"#/components/responses/Error404WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"503":{"$ref":"#/components/responses/Error503WithHeaders"}}}},"/v1/workspaces/{workspace_id}/objects/import":{"parameters":[{"name":"workspace_id","in":"path","required":true,"schema":{"$ref":"./schemas/common.v1alpha1.schema.json#/$defs/workspace_id"}}],"post":{"summary":"Snapshot explicitly allowed local files or a bundle","operationId":"importObject","tags":["objects"],"x-implementation-status":"implemented-offline","description":"Disabled unless the operator configures a named root allowed for this workspace. Requires write scope. File mode copies one relative file; bundle mode requires explicit relative includes. No absolute paths, implicit whole-root import, extraction or workload execution. Defaults: 2 GiB raw file; 100 MiB compressed and 500 MiB expanded bundle. Changed source bytes fail instead of publishing an inconsistent object. Uses the existing verified-byte and ownership-commit boundary; production SQLite/configuration composition remains deferred.","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"./schemas/object-import.v1alpha1.schema.json"}}}},"responses":{"201":{"description":"Verified bytes and workspace ownership metadata committed","headers":{"Location":{"schema":{"type":"string","format":"uri-reference"},"description":"Workspace-scoped metadata URI"},"X-Request-ID":{"$ref":"#/components/headers/RequestID"}},"content":{"application/json":{"schema":{"$ref":"./schemas/object.v1alpha1.schema.json"}}}},"400":{"$ref":"#/components/responses/Error400WithHeaders"},"401":{"$ref":"#/components/responses/Error401WithHeaders"},"403":{"$ref":"#/components/responses/Error403WithHeaders"},"404":{"$ref":"#/components/responses/Error404WithHeaders"},"408":{"$ref":"#/components/responses/Error408WithHeaders"},"413":{"$ref":"#/components/responses/Error413WithHeaders"},"415":{"$ref":"#/components/responses/Error415WithHeaders"},"429":{"$ref":"#/components/responses/Error429WithHeaders"},"503":{"$ref":"#/components/responses/Error503WithHeaders"}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"opaque workspace token"}},"schemas":{"JobSpec":{"$ref":"./schemas/job-spec.v1alpha1.schema.json"},"JobAdmission":{"$ref":"./schemas/job-admission.v1alpha1.schema.json"},"JobValidation":{"$ref":"./schemas/job-validation.v1alpha1.schema.json"},"JobStatus":{"$ref":"./schemas/job-status.v1alpha1.schema.json"},"Operation":{"$ref":"./schemas/operation.v1alpha1.schema.json"},"Error":{"$ref":"./schemas/error.v1alpha1.schema.json"},"ResultManifest":{"$ref":"./schemas/result-manifest.v1alpha1.schema.json"},"RuntimeConfig":{"$ref":"./schemas/runtime-config.v1alpha1.schema.json"},"Object":{"$ref":"./schemas/object.v1alpha1.schema.json"},"ObjectImport":{"$ref":"./schemas/object-import.v1alpha1.schema.json"},"BundleManifest":{"$ref":"./schemas/bundle-manifest.v1.schema.json"}},"headers":{"RequestID":{"schema":{"type":"string"},"description":"Server-generated request identifier; incoming IDs are not reflected"},"RetryAfter":{"schema":{"type":"string","pattern":"^[0-9]+$"},"description":"Seconds until a subsequent request can be considered"}},"responses":{"Error400WithHeaders":{"description":"Malformed request or schema violation","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error403WithHeaders":{"description":"Workspace or operation is forbidden","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error429WithHeaders":{"description":"Local admission or queue limit reached","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"},"Retry-After":{"$ref":"#/components/headers/RetryAfter"}}},"Error503WithHeaders":{"description":"Local runtime cannot safely admit work","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error401WithHeaders":{"description":"Missing or invalid runtime authentication","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error400":{"description":"Malformed request or schema violation","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error401":{"description":"Missing or invalid runtime authentication","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error403":{"description":"Workspace or operation is forbidden","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error409":{"description":"Idempotency conflict, illegal transition, or unresolved execution","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error422":{"description":"Job requirements cannot be satisfied","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error429":{"description":"Local admission or queue limit reached","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error503":{"description":"Local runtime cannot safely admit work","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error404":{"description":"Job not found or not visible","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error404Operation":{"description":"Operation not found or not visible","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}}},"Error408WithHeaders":{"description":"Request transfer did not complete within its deadline","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error413WithHeaders":{"description":"Request or object exceeds configured byte limit","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error415WithHeaders":{"description":"Unsupported or encoded content type","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}},"Error404WithHeaders":{"description":"Object not found or not visible","content":{"application/json":{"schema":{"$ref":"./schemas/error.v1alpha1.schema.json"}}},"headers":{"X-Request-ID":{"$ref":"#/components/headers/RequestID"}}}}}}